Split the pin mode machinery out and fix the wiring tab stop
